requirements.txt
fastapi==0.111.1
openai==1.37.1
python-dotenv==1.0.1
python-fasthtml==0.2.1
uvicorn==0.30.3